+++ src/server/pkg/media/types_test.go
@@ -1,4 +1,4 @@
-package files
+package media
 
 import (
 	"encoding/json"
